% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.scalar.R
\name{is.scalar}
\alias{is.scalar}
\title{is.scalar}
\usage{
is.scalar(x)
}
\arguments{
\item{x}{input.}
}
\value{
TRUE if x is a scalar or FAULSE otherwise.
}
\description{
Check if the input is a scalar.
}
\examples{
is.scalar(100)
is.scalar(1.0259e+3)
is.scalar(seq(from = 0.0, to = 10, by = 1))
is.scalar(t(cbind(c(0.0, 0.0, -5.0), c(1.0, 0.0, -5.0), c(2.0, 0.0, -5.0), c(3.0, 0.0, -5.0), c(4.0, 0.0, -5.0), c(5.0, 0.0, -5.0), c(6.0, 0.0, -5.0), c(7.0, 0.0, -5.0), c(8.0, 0.0, -5.0), c(9.0, 0.0, -5.0), c(10.0, 0.0, -5.0))))

}
