% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavcomp.R
\name{wavcomp}
\alias{wavcomp}
\title{wavcomp}
\usage{
wavcomp(
  h,
  g,
  n,
  m,
  freq,
  kata,
  spevar,
  dfreq = NULL,
  dkata = NULL,
  imodel = 1,
  iseed = 150001,
  tol = 1e-05,
  no = 200
)
}
\arguments{
\item{h}{Water depth (from MWL to ocean bottom), m or ft.}

\item{g}{Gravitational acceleration, m/s^2 or ft/s^2.}

\item{n}{Total number of wave frequencies in a directional wave energy spectrum.}

\item{m}{Total number of wave directions in a directional wave energy spectrum.}

\item{freq}{Wave frequency scalar or vector [n], Hz.}

\item{kata}{Wave direction scalar or vector [m], degree.}

\item{spevar}{Directional wave spectral variance array [m x n], m^2/Hz/deg or ft^2/Hz/deg.}

\item{dfreq}{(Optional, default NULL) If NULL, wave frequency band width vector dfreq shall be calculated based on the provided wave frequency vector freq; if provided, the length of dfreq shall be 1 or n. For n = 1, dfreq must be given.}

\item{dkata}{(Optional, default NULL) If NULL, wave direction band width vector dkata shall be calculated based on the provided wave direction vector kata; if provided, the length of dkata shall be 1 or m. For m = 1, dkata must be given.}

\item{imodel}{(Optional, default imodel = 1) Choose linear superposition model for directional wave simulation: 1 for Single Summation Model, 2 for Single Summation Model.}

\item{iseed}{(Optional, default iseed = 150001) Initial large odd integer for generating randomly selected initial phases.}

\item{tol}{(Optional, default tol = 1.0e-5) Acceptance tolerance of convergence test for Newton-Raphson Method.}

\item{no}{(Optional, default no = 200) Maximum number of iteration for Newton-Raphson Method.}
}
\value{
Return regular wave properties such as amplitude, wave number, wave circular frequency, wave direction, and wave initial random phase for each wave components.
}
\description{
Obtain wave properties, such as wave amplitude, wave number, wave circular frequency,
wave direction, and wave initial random phase for each regular wave component from a
directional wave energy spectrum which describes the energy content of an ocean wave
and its distribution over a frequency range and a direction range of the random wave.
This function is called by DWS.
}
\examples{
wavcomp(h = 100.0, g = 9.80665, n = 1, m = 1, freq = 0.1, kata = 45, spevar = 4.5, dfreq = 1, dkata = 1)
wavcomp(h = 100.0, g = 9.80665, n = 2, m = 1, freq = c(0.1, 0.3), kata = 45, spevar = matrix(c(4.5, 0.5), 1, 2), dfreq = 1, dkata = 1)
wavcomp(h = 100.0, g = 9.80665, n = 1, m = 2, freq = 0.1, kata = c(0, 45), spevar = matrix(c(0.5, 4.5), 2, 1), dfreq = 1, dkata = 1)
wavcomp(h = 100.0, g = 9.80665, n = 2, m = 2, freq = c(0.1, 0.3), kata = c(0, 45), spevar = matrix(c(0.0, 4.5, 0.5, 0.5), 2, 2), dfreq = c(rep(1, 2)), dkata = c(rep(1, 2)))
wavcomp(h = 100.0, g = 9.80665, n = 2, m = 2, freq = c(0.1, 0.3), kata = c(0, 45), spevar = matrix(c(0.0, 4.5, 0.5, 0.5), 2, 2), dfreq = c(rep(1, 2)), dkata = c(rep(1, 2)), imodel = 1, iseed = 150001, tol = 1.0e-5, no = 200)

}
