% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavnum.R
\name{wavnum}
\alias{wavnum}
\title{wavnum}
\usage{
wavnum(h, g, sigma, tol = 1e-05, no = 200)
}
\arguments{
\item{h}{Water depth, m or ft}

\item{g}{Gravitational acceleration, m/s^2 or ft/s^2}

\item{sigma}{Circular wave frequency, rad/s}

\item{tol}{(Optional,default tol = 1.0e-5). Acceptance tolerance of convergence test for Newton-Raphson Method.}

\item{no}{(Optional, default no = 200) Maximum number of iteration for Newton-Raphson Method.}
}
\value{
An approximate solution wave number dk or error message of convergence failure to the console.
}
\description{
Calculate wave number from the Linear Dispersion Relation using Newton-Raphson Method.
This function is called by function wavcomp.
}
\examples{
wavnum(h = 100.0, g = 9.80665, sigma = 0.1000)
wavnum(h = 100.0, g = 9.80665, sigma = 0.1000, tol = 1.0e-5)
wavnum(h = 100.0, g = 9.80665, sigma = 0.1000, tol = 1.0e-5, no = 200)

}
