% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavsimu.R
\name{wavsimu}
\alias{wavsimu}
\title{wavsimu}
\usage{
wavsimu(h, g, rho, x, y, z, t, amp, fk, fsigma, fkata, beta, itheory = 1)
}
\arguments{
\item{h}{Water depth (from MWL to ocean bottom), m or ft.}

\item{g}{Gravitational acceleration, m/s^2 or ft/s^2.}

\item{rho}{Ocean water density, kg/m^3 or slugs/ft^3.}

\item{x}{X-coordinate of a specified location (positive in the wave direction of 0 deg), m or ft.}

\item{y}{Y-coordinate to a specified ocation (positive in the wave direction of 90 deg), m or ft.}

\item{z}{Z-coordinate to a specified location (positive upwards, zero at MWL), m or ft.}

\item{t}{Time instance to calculate water surface displacement and velocity potential at location (x,y,z), sec}

\item{amp}{Regular wave amplitude vector for each wave components}

\item{fk}{Regular wave number vector for each wave components}

\item{fsigma}{Regular wave circular frequency vector for each wave components}

\item{fkata}{Regular wave direction vector for each wave components}

\item{beta}{Regular wave initial random phase vector for each wave components}

\item{itheory}{(Optional, default itheory = 1) Choose wave theory: 1 for Linear Wave Theory, 2 for Linear Wave Theory with Wheeler's Stretching.}
}
\value{
A time instance of total irregular wave potential, total irregular wave horizontal X-velocity, total irregular wave horizontal Y-velocity, total irregular wave horizontal Z-velocity, total irregular wave horizontal X-acceleration, total irregular wave horizontal Y-acceleration, total irregular wave vertical Z-acceleration, and total irregular wave dynamic pressure at location (x, y, z)
}
\description{
Construct total irregular wave at location (x, y, z) at time instance t by linear superposition
of all regular wave components obtained from a directional wave energy spectrum.
This function is called by DWS.
}
\examples{
wavsimu(h = 100.0, g = 9.80665, rho = 1030.0, x = 0.0, y = 0.0, z = -5.0, t = 0.0, amp = 3, fk = 0.04028232, fsigma = 0.6283185, fkata = 0.7853982, beta = 0.1369422)
wavsimu(h = 100.0, g = 9.80665, rho = 1030.0, x = 0.0, y = 0.0, z = -5.0, t = 0.0, amp = 3, fk = 0.04028232, fsigma = 0.6283185, fkata = 0.7853982, beta = 0.1369422, itheory = 2)
wavsimu(h = 100.0, g = 9.80665, rho = 1030.0, x = 0.0, y = 0.0, z = -5.0, t = 0.0, amp = c(3, 1), fk = c(0.04028293, 0.36231658), fsigma = c(0.6283185, 1.884956), fkata = c(0.7853982, 0.7853982), beta = c(0.1369422, 1.0299106))
wavsimu(h = 100.0, g = 9.80665, rho = 1030.0, x = 0.0, y = 0.0, z = -5.0, t = 0.0, amp = c(0, 3, 1, 1), fk = c(0.04028293, 0.04028694, 0.36231658, 0.36232866), fsigma = c(0.6283185, 0.6283499, 1.8849556, 1.8849870), fkata = c(0.0000000, 0.7853982, 0.0000000, 0.7853982), beta = c(0.1369422, 1.0299106, 4.0621247, 5.2208995))

}
