% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavspec.R
\name{wavspec}
\alias{wavspec}
\title{wavspec}
\usage{
wavspec(spectrum, dfreq = NULL, dkata = NULL)
}
\arguments{
\item{spectrum}{A matrix (at lease 2 X 2) contains all information of a directional wave energy spectrum.}

\item{dfreq}{(Optional, default NULL) Wave frequency band width scalar or vector.}

\item{dkata}{(Optional, default NULL) Wave direction band width scalar or vector.}
}
\value{
All necessary wave spectral information for function DWS.
}
\description{
Obtain spectral information from a directional wave energy spectrum.
}
\examples{
wavspec(spectrum)
wavspec(spectrum, dfreq = 1, dkata = 1)
wavspec(spectrum, dfreq = c(rep(1, 23)), dkata = c(rep(1, 17)))

}
