% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavtsplot.R
\name{wavtsplot}
\alias{wavtsplot}
\title{wavtsplot}
\usage{
wavtsplot(LC, iunit = 1)
}
\arguments{
\item{LC}{Output list from DWS function.}

\item{iunit}{(Optional, default iunit = 1) Choose unit system: 1 for SI unit system, 2 for English Unit.}
}
\value{
A plotting window with 3 X 3 matrix for time series plots of wave properties.
}
\description{
Plot time series of wave properties in a plotting window with 3 X 3 matrix based on a output list from DWS function.
}
\examples{
wavtsplot(LC1)
wavtsplot(LC = LC1, iunit = 1)

}
